requirements.txt
absl-py==1.4.0
abstract_singleton==1.0.1
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
astunparse==1.6.3
async-generator==1.10
async-timeout==4.0.2
asynctest==0.13.0
attrs==22.2.0
auto_gpt_plugin_template==0.0.2
autoflake==2.1.1
autopep8==2.0.2
beautifulsoup4==4.12.2
black==23.3.0
blis==0.7.9
cachetools==5.3.0
catalogue==2.0.8
certifi==2022.12.7
cffi==1.15.1
cfgv==3.3.1
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==2.2.1
colorama==0.4.6
confection==0.0.4
contourpy==1.0.7
coverage==7.2.3
cssselect==1.2.0
cycler==0.11.0
cymem==2.0.7
discord.py==2.2.2
distlib==0.3.6
dnspython==2.3.0
docker==6.0.1
duckduckgo-search==2.8.6
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.4.0/en_core_web_sm-3.4.0-py3-none-any.whl#sha256=35365d49d389ecb19ba33119dca9ab9ff80b36fbeb063fa5d44008197fead8fa
exceptiongroup==1.1.1
filelock==3.11.0
flake8==6.0.0
flatbuffers==23.3.3
fonttools==4.39.3
frozenlist==1.3.3
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.31
google-api-core==2.11.0
google-api-python-client==2.86.0
google-auth==2.17.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
googleapis-common-protos==1.59.0
grpcio==1.53.0
gTTS==2.3.1
gym==0.26.2
gym-notices==0.0.8
h11==0.14.0
h5py==3.8.0
httpcore==0.17.0
httplib2==0.22.0
httpx==0.24.0
identify==2.5.22
idna==3.4
iniconfig==2.0.0
isort==5.12.0
jax==0.4.8
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
keras==2.12.0
kiwisolver==1.4.4
langcodes==3.3.0
libclang==16.0.0
loguru==0.7.0
lxml==4.9.2
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
mccabe==0.7.0
ml-dtypes==0.1.0
mpmath==1.3.0
multidict==6.0.4
murmurhash==1.0.9
mypy-extensions==1.0.0
networkx==3.1
nodeenv==1.7.0
numpy==1.23.5
oauthlib==3.2.2
openai==0.27.2
openapi-python-client==0.13.4
opt-einsum==3.3.0
orjson==3.8.10
outcome==1.2.0
packaging==23.1
pathspec==0.11.1
pathy==0.10.1
Pillow==9.5.0
pinecone-client==2.2.1
platformdirs==3.2.0
playsound==1.2.2
pluggy==1.0.0
pre-commit==3.2.2
preshed==3.0.8
protobuf==4.22.3
py-cpuinfo==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycparser==2.21
pydantic==1.10.7
pyflakes==3.0.1
pyglet==1.5.21
pyparsing==3.0.9
pyrsistent==0.19.3
PySocks==1.7.1
pytest==7.3.1
pytest-asyncio==0.21.0
pytest-benchmark==4.0.0
pytest-cov==4.0.0
pytest-integration==0.2.3
pytest-mock==3.10.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pywin32==306
PyYAML==6.0
readability-lxml==0.8.1
redis==4.5.4
regex==2023.3.23
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
selenium==4.8.3
shellingham==1.5.0.post1
six==1.16.0
smart-open==6.3.0
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
spacy==3.4.4
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.6
sympy==1.11.1
tensorboard==2.12.2
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.2.0
thinc==8.1.9
threadpoolctl==3.1.0
tiktoken==0.3.3
torch==2.0.0
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.2
tweepy==4.13.0
typer==0.7.0
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.15
virtualenv==20.22.0
wasabi==0.10.1
webdriver-manager==3.8.6
websocket-client==1.5.1
Werkzeug==2.2.3
win32-setctime==1.1.0
wrapt==1.14.1
wsproto==1.2.0
yarl==1.8.2
mss==6.1.0
pytesseract
